linked editor and plotter.
characters.fs - the file where we will be writing the words to instruct the plotter to draw each characeter. currently the plot-"letter" words just emit the character to simulate what the plotter would do. i did this just for a sanity check on the word logic.

shapes.fs and writing.fs will be for a single character. these will be implemented on a different branch.

moving forward here are my goals:
- revamp the editor UI
- add backspaces and cursor movement?